# source/api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is Process Wiz App.",
    "version": "1.0.0",
    "title": "Process Wiz"
  },
  "host": "api.pwdev.internal.nlightnconsulting.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/doc-guide/chat": {
      "post": {
        "tags": [
          "DocGuide"
        ],
        "description": "API query doc guide chat.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc guide chat.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DocGuideRequestObject"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/doc-guide/sections/chat": {
      "post": {
        "tags": [
          "DocGuide"
        ],
        "description": "API to query doc guide sections.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to send for doc guide chat.",
            "required": true,
            "schema": {
              "properties": {
                "user_query": {
                  "type": "string"
                },
                "file_id": {
                  "type": "number"
                },
                "section_id": {
                  "type": "number"
                },
                "thread_id": {
                  "type": "string"
                },
                "stage_id": {
                  "type": "string",
                  "enum": [
                    "INITIAL_CHAT",
                    "CONVERSATION",
                    "TIMED_OUT",
                    "COMPLETED"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/doc-guide/files": {
      "get": {
        "tags": [
          "DocGuide"
        ],
        "description": "API to list files in doc guide.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DocGuideFileResponse"
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/doc-guide/collection-files": {
      "get": {
        "tags": [
          "DocGuide"
        ],
        "description": "API to list common files for doc guide process.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/CollectionDocGuideResponse"
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/doc-guide/sections": {
      "get": {
        "tags": [
          "DocGuide"
        ],
        "description": "API to list sections in doc guide.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "file_id",
            "description": "Id of the file which the sections contain.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DocGuideSectionResponse"
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Type in the *'Value'* input box below: **'Bearer &lt;JWT&gt;'**, where JWT is the token"
    },
    "organization-id": {
      "type": "apiKey",
      "in": "header",
      "name": "organization-id",
      "description": "Id of the organization"
    }
  },
  "definitions": {
...
    "DocFile": {
      "type": "object",
      "properties": {
        "file_id": {
          "type": "number"
        },
        "file_name": {
          "type": "string"
        },
        "file_url": {
          "type": "string"
        },
        "is_processed": {
          "type": "boolean"
        }
      }
    },
...
    "CollectionDocGuide": {
      "type": "object",
      "properties": {
        "collection_id": {
          "type": "number"
        },
        "collection_name": {
          "type": "string"
        },
        "collection_description": {
          "type": "string"
        },
        "files": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DocFile"
          }
        }
      }
    },
...
    "CollectionDocGuideResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CollectionDocGuide"
          }
        }
      }
    },
...
    "NormalFailureResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "DocGuideRequestObject": {
      "type": "object",
      "properties": {
        "user_query": {
          "type": "string"
        },
        "file_id": {
          "type": "integer"
        },
        "thread_id": {
          "type": "string"
        }
      }
    },
...
    "DocGuideFileResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DocGuideFile"
          }
        }
      }
    },
    "DocGuideFile": {
      "type": "object",
      "properties": {
        "file_id": {
          "type": "number"
        },
        "file_name": {
          "type": "string"
        },
        "file_url": {
          "type": "string"
        }
      }
    },
    "DocGuideSectionResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DocGuideSection"
          }
        }
      }
    },
    "DocGuideSection": {
      "type": "object",
      "properties": {
        "file_id": {
          "type": "integer"
        },
        "section_id": {
          "type": "number"
        },
        "section_title": {
          "type": "string"
        },
        "section_status": {
          "type": "string"
        },
        "section_assessment_score": {
          "type": "integer"
        },
        "is_enabled": {
          "type": "boolean"
        },
        "latest_thread_id": {
          "type": "string"
        }
      }
    }
  }
}